* Remove remark about wrapping Selenium::Remote::Driver
Because that's explicitly *not* what this module tries to achieve! The
module does provide a DriverRole role for which there is a wrapper for
Selenium::Remote::Driver. The point of this module is to bridge the gap
between the PageObject approach of modelling web page interaction and
the *various* low-level web driving technologies (of which Selenium is
just an example/ variant).